make InputStream used in S3Storage object
