+++ sbt_common/storage/src/main/scala/uk/ac/wellcome/storage/s3/S3Storage.scala
@@ -1,0 +1,51 @@
+package uk.ac.wellcome.storage.s3
+
+import java.io.InputStream
+
+import com.amazonaws.services.s3.AmazonS3
+import com.twitter.inject.Logging
+
+import scala.concurrent.Future
+import scala.util.hashing.MurmurHash3
+
+object S3Storage extends Logging {
+  def put(s3Client: AmazonS3, bucketName: String)(keyPrefix: String)(
+    is: InputStream): Future[S3ObjectLocation] = {
+
+    // Currently hiding the stringification, so we can deal with it later if we need to
+    val content = scala.io.Source.fromInputStream(is).mkString
+    val contentHash = MurmurHash3.stringHash(content, MurmurHash3.stringSeed)
+
+    // Ensure that keyPrefix here is normalised for concatenating with contentHash
+    val prefix = keyPrefix
+      .stripPrefix("/")
+      .stripSuffix("/")
+
+    val key = s"$prefix/$contentHash.json"
+
+    info(s"Attempt: PUT object to s3://$bucketName/$key")
+    val putObject = Future {
+      s3Client.putObject(bucketName, key, content)
+    }
+
+    putObject.map { _ =>
+      info(s"Success: PUT object to s3://$bucketName/$key")
+      S3ObjectLocation(bucketName, key)
+    }
+  }
+
+  def get(s3Client: AmazonS3, bucketName: String)(
+    key: String): Future[InputStream] = {
+    info(s"Attempt: GET object from s3://$bucketName/$key")
+
+    val futureInputStream = Future {
+      s3Client.getObject(bucketName, key).getObjectContent
+    }
+
+    futureInputStream.foreach { case _ =>
+      info(s"Success: GET object from s3://$bucketName/$key")
+    }
+
+    futureInputStream
+  }
+}
